Docker build files for Oracle Database (#118)
* Container starting successfully

* Oracle DB EE working

* Including build scrip

buildDockerImage.sh can now be used to drive a Docker build. It also
includes checksum verification of the downloaded files.

* Enable PDB OPEN at startup

* Fixed LISTENER unknown service

* Changes in README.md

* Fixed broken link

* Cleanup install binaries right after install

* Removed yum upgrade

* Provide admin account pwd, rename image

It is now possible to provide a password for the admin accounts for the
Database (SYS, SYSTEM, PDBAMIN). In case that no password has been
specified the build script will generate one.

The image has been renamed to oracle/database:<version>-<edition>

* Update README.md with running instructions

* Fixed typo

* Include doc and steps on how to connect

* Include OEM URL

* Update README.md

* Update buildDockerImage.sh

* Update Dockerfile.ee

* Decrease layers in docker file