12212-domain: Make user setting Weblogic password successfully
While specifying the "--build-arg ADMIN_PASSWORD=welcome1" option for
building image the varible ADMIN_PASSWORD defined in Dockerfile will
never be set due to Docker known mechanism block this behavior. It
causes the image will not be built successfully, so this patch intends
to fix this issue. User can specify the password or not. If not the
building process will generate the password automatically.

Signed-off-by: Dong Zhu <dong.zhu@oracle.com>