Updated dockerfiles according to the comments of PR495
1. Updated Dockerfiles, init.sh and README.md according to the comments.
2. Deleted the unnecessary templates.
3. For the comment, "This USER directive is not necessary as you're
already root", we tried to delete the changing user to root, but failed
to install softwares using yum install XXX, it says it's not root. The
reason might be that the USER has been changed to ORACLE in the base
image oracle/tuxedo:12.2.2. So we did not delete the directive.