managed server listen address is always set to IP (#568)
* managed server listen address is always set to IP

When a managed server is created, the listen address is always set to IP
address.  This occurs even if the MS_HOST variable is passed.  The
behavior can be observed in the WebLogic console.  A problem can arise
if the managed server container is restarted and it does not get the
same ip.  The result is the managed server will not start.  To fix this
the and add-server.py script is updated to set listen address as
hostname when the MS_HOST variable is passed.  (the host name should be
passed as the container name).  The container IP will be resolved as
long as the default docker bridge is not being used.