Read JAVA_OPTIONS fix
If you add JVM arguments of type -Dkey1=value1 -D key2=value2... in JAVA_OPTIONS property of file security.properties, only the "-Dkey1" part gets captured and added as JVM arguments when starting the server. Adding the "-" to the "-f2" argument fix the problem because you get the rest of the value.