Changes for spike troubleshooting
Removed OSB from image as it was too large.
Also skipped RCU as once run in the database it was then erroring as not
detecting that it had been run and was trying to run again.

Now failing after trying to set up domain (again this has already been
done on DB so failing now) so sleep added to prevent container from
restarting instantly.