[docs] The org.opencontainers.image labels need to be mandatory
The oraclelinux:7-slim and oraclelinux:8-slim images are now
shipping with these labels. If a product does not provide more
specific values, then it will appear to products that consume
these labels as "Oracle Linux" and not a more specific product image.

This is being noticed in particular for images that are published to
GitHub Container Registry that does consume these tags.

Signed-off-by: Avi Miller <avi.miller@oracle.com>