feat(dev-images): switch the Oracle Linux Developer image base
This change modifies the base image used by the Oracle Linux 8
based images from the slim variant to the full variant for better
developer alignment, i.e. availability of the dnf tool.

An intermediate compat image is used to install microdnf so that
existing tooling for these images doesn't break.

Signed-off-by: Avi Miller <avi.miller@oracle.com>