feat(developper): :sparkles: add golang 1.16 container (#2085)
Add a a golang 1.16 specific container. For now it will generate the
same container as the one in the `golang` directory, but it "locks" the
version on 1.16.
It will also be built by the `build-and-push-dev-images` action.
Also update containers from previous golang version to disable any
pre-set golang channel. This will avoid issue if/when 1.16 is no more
the default version.

Signed-off-by: Philippe Vanhaesendonck <philippe.vanhaesendonck@oracle.com>