Install DB and apply patches in a single layer
The oracle-docker-images repository allows patches to be applied, but the resulting image size is doubled (see https://github.com/oracle/docker-images/issues/1990). Instead, we would like to apply the patches in the same layer where the database binaries are installed.

I have applied this only to the 19c and 21c Dockerfiles since the other versions are either at or rapidly approaching end of life.

Changes:
- Set SLIMMING to false to ensure that all files required for patching are present.
- Copied patchDBBinaries.sh, runDatapatch.sh and savePatchSummary.sh from the extensions/patch directory.
- Modified patchDBBinaries.sh to delete temporary files to reduce image size.
- Added steps to builder stage in Dockerfile to copy the patches to the container and run the patchDBBinaries.sh script.
- Added steps to Dockerfile to copy datapatch and lspatches scripts under $ORACLE_BASE.